Database test
Here is the project for my database test. It uses a database to
dynamically create a dropbox with all of the entries, then uses the
selected entry to populate the response page. I'm still not sure how to
put stuff into the database, but at least we can access stuff from it.
This might not work if you don't have the same DB setup on your machine.